+++ pyfda/input_widgets/input_pz_ui.py
@@ -17,7 +17,7 @@ from pyfda.libs.compat import (
     QRadioButton, QFrame, QSpinBox, QFont, QIcon, QVBoxLayout, QHBoxLayout)
 from pyfda.libs.pyfda_qt_lib import qstyle_widget, qcmb_box_populate, emit
 from pyfda.libs.pyfda_qt_classes import PushButton
-from pyfda.libs.csv_option_box import CSV_option_box
+from pyfda.libs.csv_option_box import CSVOptionBox
 from pyfda.libs.pyfda_lib import to_html, first_item
 import pyfda.libs.pyfda_dirs as dirs
 from pyfda.pyfda_rc import params
@@ -315,7 +315,7 @@ class Input_PZ_UI(QWidget):
             if self.but_csv_options.isChecked():
                 # Important: Handle to window must be class attribute otherwise it (and
                 # the attached window) is deleted immediately when it goes out of scope
-                dirs.csv_options_handle = CSV_option_box()
+                dirs.csv_options_handle = CSVOptionBox()
                 dirs.csv_options_handle.sig_tx.connect(self.process_sig_rx)
                 dirs.csv_options_handle.show()  # modeless i.e. non-blocking popup window
                 # alert other widgets that csv options / visibility have changed